when you player select get player
