+++ client/src/App.tsx
@@ -1,18 +1,15 @@
 import { useEffect } from 'react'
 import axios from 'axios'
 import axiosRetry from 'axios-retry'
-import { ResPlayer, ResEnemies, ResCard } from './types/api/response'
-import { useAppSelector, useAppDispatch } from './redux/hooks'
-import { setPlayer, initialDeck } from './redux/slice/playerSlice'
+import { ResEnemies, ResCard } from './types/api/response'
+import { useAppDispatch } from './redux/hooks'
 import { setEnemies } from './redux/slice/enemiesSlice'
 import { setCards } from './redux/slice/cardsSlice'
 import GameTitle from './components/gameTitle'
 import RootSelect from './components/rootSelect'
 import Battle from './components/battle'
-import { initializeDeck } from './battle/deck'
 
 const App = (): JSX.Element => {
-  const cards = useAppSelector((state) => state.cards)
   const dispatch = useAppDispatch()
 
   const axiosClient = axios.create({ baseURL: process.env.REACT_APP_API_URL_BROWSER })
@@ -29,17 +26,6 @@ const App = (): JSX.Element => {
     })
   }
 
-  const getPlayer = async (): Promise<void> => {
-    await axiosClient.get('/v1/players')
-    .then(res => {
-      const resPlayer: ResPlayer = res.data
-      dispatch(setPlayer(resPlayer))
-    })
-    .catch(error => {
-      console.log("プレイヤーの取得に失敗しました")
-    })
-  }
-
   const getCards = async (): Promise<void> => {
     await axiosClient.get('/v1/cards')
     .then(res => {
@@ -53,14 +39,8 @@ const App = (): JSX.Element => {
 
   useEffect((): void => {
     getEnemies()
-    getPlayer()
     getCards()
   }, [])
-
-  useEffect((): void => {
-    const defaultDeck = initializeDeck(cards)
-    dispatch(initialDeck(defaultDeck))
-  }, [cards])
 
   return (
     <div>
